+++ src/types/account.ts
@@ -13,14 +13,3 @@ export type Secretkey = {
   secretkeyType: SecretKeyType;
   secretkey: string;
 };
-
-export type KeyType = 'mnemonic' | 'privatekey' | 'keystore';
-
-export type AddressListItem = {
-  label: string;
-  keyType: KeyType;
-  address: string;
-  selected?: boolean;
-};
-
-export type AddressList = AddressListItem[];
